This is still only a rough implementation but from the current logic assets are created in a single place before the program is executed / bundled.
(this place would be src/init/assets/index.mts -> initializeAssets).

Assets that are not used in the source folder (src/) will not be included.

This is useful if you have a bunch of typescript files but you only
need to import one of them (like here with the runtime implementation).

This would also make it possible to request different "versions" of the
same asset through a different protocol like so:

text://my-typescript-file.mts would return the asset in its raw text form
js-bundle://my-typescript-file.mts would return a javascript bundle

Etc.

It may be worth while looking into what assets are needed by the different
exports since this list (of used assets) is used for *all* exports.